+++ src/Html.php
@@ -19,6 +19,7 @@ use Yiisoft\Html\Tag\CustomTag;
 use Yiisoft\Html\Tag\Datalist;
 use Yiisoft\Html\Tag\Div;
 use Yiisoft\Html\Tag\Em;
+use Yiisoft\Html\Tag\Fieldset;
 use Yiisoft\Html\Tag\Form;
 use Yiisoft\Html\Tag\H1;
 use Yiisoft\Html\Tag\H2;
@@ -34,6 +35,7 @@ use Yiisoft\Html\Tag\Input\File;
 use Yiisoft\Html\Tag\Input\Radio;
 use Yiisoft\Html\Tag\Input\Range;
 use Yiisoft\Html\Tag\Label;
+use Yiisoft\Html\Tag\Legend;
 use Yiisoft\Html\Tag\Li;
 use Yiisoft\Html\Tag\Link;
 use Yiisoft\Html\Tag\Audio;
@@ -535,6 +537,20 @@ final class Html
     }
 
     /**
+     * Generates a {@see Fieldset} tag.
+     *
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     */
+    public static function fieldset(array $attributes = []): Fieldset
+    {
+        $tag = Fieldset::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $tag;
+    }
+
+    /**
      * Generates a {@see Form} tag.
      *
      * @param string|null $action The URL to use for form submission.
@@ -571,6 +587,24 @@ final class Html
         }
         if ($content !== '') {
             $tag = $tag->content($content);
+        }
+        return $tag;
+    }
+
+    /**
+     * Generates a {@see Legend} tag.
+     *
+     * @param string|Stringable $content The tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     */
+    public static function legend($content = '', array $attributes = []): Legend
+    {
+        $tag = Legend::tag();
+        if ($content !== '') {
+            $tag = $tag->content($content);
+        }
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
         }
         return $tag;
     }
@@ -633,8 +667,9 @@ final class Html
      * @param string|null $name The name attribute. If it is `null`, the name attribute will not be generated.
      * @param bool|float|int|string|Stringable|null $value The value attribute. If it is `null`, the value
      * attribute will not be generated.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
      */
-    public static function input(string $type, ?string $name = null, $value = null): Input
+    public static function input(string $type, ?string $name = null, $value = null, array $attributes = []): Input
     {
         $tag = Input::tag()->type($type);
         if ($name !== null) {
@@ -642,6 +677,9 @@ final class Html
         }
         if ($value !== null) {
             $tag = $tag->value($value);
+        }
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
         }
         return $tag;
     }
@@ -652,10 +690,15 @@ final class Html
      * @see Input::button()
      *
      * @param string|null $label The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
      */
-    public static function buttonInput(?string $label = 'Button'): Input
+    public static function buttonInput(?string $label = 'Button', array $attributes = []): Input
     {
-        return Input::button($label);
+        $tag = Input::button($label);
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $tag;
     }
 
     /**
@@ -664,10 +707,15 @@ final class Html
      * @see Input::submitButton()
      *
      * @param string|null $label The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
      */
-    public static function submitInput(?string $label = 'Submit'): Input
+    public static function submitInput(?string $label = 'Submit', array $attributes = []): Input
     {
-        return Input::submitButton($label);
+        $tag = Input::submitButton($label);
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $tag;
     }
 
     /**
@@ -676,10 +724,15 @@ final class Html
      * @see Input::resetButton()
      *
      * @param string|null $label The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
      */
-    public static function resetInput(?string $label = 'Reset'): Input
+    public static function resetInput(?string $label = 'Reset', array $attributes = []): Input
     {
-        return Input::resetButton($label);
+        $tag = Input::resetButton($label);
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $tag;
     }
 
     /**