+++ src/Widget/RadioList/RadioList.php
@@ -10,7 +10,7 @@ use Yiisoft\Html\NoEncodeStringableInterface;
 use Yiisoft\Html\Tag\Input;
 
 /**
- * RadioList represents a list of radios and their corresponding labels.
+ * `RadioList` represents a list of radios and their corresponding labels.
  */
 final class RadioList implements NoEncodeStringableInterface
 {
@@ -32,9 +32,6 @@ final class RadioList implements NoEncodeStringableInterface
      */
     private array $items = [];
 
-    /**
-     * @psalm-param non-empty-string
-     */
     private string $name;
 
     private ?string $value = null;
@@ -52,6 +49,13 @@ final class RadioList implements NoEncodeStringableInterface
     public static function create(string $name): self
     {
         return new self($name);
+    }
+
+    public function name(string $name): self
+    {
+        $new = clone $this;
+        $new->name = $name;
+        return $new;
     }
 
     public function withoutContainer(): self