+++ CHANGELOG.md
@@ -7,9 +7,15 @@
 - Chg #105: Rename `Html::fileInput()` to `Html::file()`, `Html::file()` and `Input::file()` returns `File` (vjik)
 - New #109: Add class for tag `Datalist` and method `Html::datalist()` (vjik)
 - New #109: Add specialized class for input tag with type `Range` and methods `Html::range()`, `Input::range()` (vjik)
+- New #113: Add class for tag `Legend`, class for tag `Fieldset`, methods `Html::legend()` and `Html::fieldset()` (vjik)
+- New #111: Add widget `ButtonGroup` (vjik)
+- New #111: Add method `Tag::unionAttributes()` that available for all tags (vjik)
 - Enh #106: Add option groups support to method `Select::optionsData()` (vjik)
 - Enh #108: Add individual attributes of options and option groups support to method `Select::optionsData()` (vjik)
 - Enh #102: Remove psalm type `HtmlAttributes`, too obsessive for package users (vjik)
+- Enh #104: Add parameter `$attributes` to methods `Html::input()`, `Html::buttonInput()`, `Html::submitInput()` 
+  and `Html::resetInput()` (vjik)
+- Enh #115: Add methods `CheckboxList::name()` and `RadioList::name()` (vjik)
 
 ## 2.3.0 March 25, 2022